requirements.txt
python-docx>=0.8.11
openpyxl>=3.0.10
PyQt6>=6.4.0
PyInstaller>=5.13.0